Cleaning up fields test
