--- test/fields.js
@@ -1,1 +1,0 @@
-require('./_fields-helper').init(true);
